Rearrange entry details layout
