+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.8.121+1211
+version: 0.8.122+1212
 
 msix_config:
   display_name: Lotti
